Backport of: ------------------------------------------------------------ revno: 2476.1116.1 committer: davi@mysql.com/endora.local timestamp: Fri 2007-12-14 10:10:19 -0200 message: DROP TABLE under LOCK TABLES simultaneous to a FLUSH TABLES WITH READ LOCK (global read lock) can lead to a deadlock.
The solution is to not wait for the global read lock if the
thread is holding any locked tables.

Related to bugs 23713 and 32395. This issues is being fixed
only on 6.0 because it depends on the fix for bug 25858 --
which was fixed only on 6.0.

--BZR--
revision-id: kostja@sun.com-20091120201257-yy7t8dgotsotg0so
property-branch-nick: next-mr-kostja
testament3-sha1: 4c8820661f7d9a7cfdce3543a78b76c140cca95e