BUG#48506 crash in CREATE TABLE IF NOT EXISTS <existing_view> LIKE <tmp_tbl> with RBL
When binlogging the statement, the server always handle the existing
object as a table, even though it is a view. However a view is
handled differently in other parts of the code thus leading the
statement to crash in RBL if the view exists.

This happens because the underlying tables for the view are not opened
when we try to call store_create_info() on the view in order to build
a CREATE TABLE statement.

This patch will only address the crash problem, other binlogging
problems related to CREATE TABLE IF NOT EXISTS LIKE when the existing
object is a view will be solved by BUG 47442.
--BZR--
revision-id: alfranio.correia@sun.com-20091127133439-u24wzytx9icoqcmn
property-branch-nick: mysql-5.1-bugteam
testament3-sha1: 772d80e79ec29edab945dd6e7b7f7cdee49f7295