Backport of revno: 3711.1.1
Bug #48725 Assert !thd->is_error() in delayed_get_table()

This bug is a regression introduced by the patch for Bug #45949.

If the handler thread for INSERT DELAYED was killed by e.g.
FLUSH TABLES, the error message is copied from the handler thread
to the connection thread. But the error was not reacted on, so the
connection thread continued as normal, leading to an eventual assert.

No test case added as it would have required sync points to work
for handler threads. The plan is to add this in the scope of 
Bug #48725 / Bug #48541. The patch has been tested with the 
non-deterministic test case given in the bug description.

--BZR--
revision-id: jon.hauglid@sun.com-20091209115852-rcxh5v6dox6pbw79
property-branch-nick: mysql-next-4284-bp1
testament3-sha1: b65d9c284ecfafd73f5913aeeb18ca73df0a0195