Bug #48742  	Replication: incorrect help text for --init-slave
The help text for --init-slave=name:
"Command(s) that are executed when a slave connects to this master".
This text indicate that the --init-slave option is set on a  master 
server, and the master server passes the option's argument to slave 
which connects to it. This is wrong. Actually the --init-slave option 
just can be set on a slave server, and then the slave server executes 
the argument each time the SQL thread starts.

Correct the help text for --init-slave option as following:
"Command(s) that are executed by a slave server each time the SQL thread starts."

--BZR--
revision-id: dao-gang.qu@sun.com-20091211015738-nst7dehuyhgztvmy
property-branch-nick: mysql-5.0-bugteam
property-file-info: ld7:file_id62:sp1f-mysqld.cc-19700101030959-zpswdvekpvixxzxf7gdtofzel7nywtfj7:message46:Correct the help text for --init-slave option.4:path13:sql/mysqld.ccee
testament3-sha1: 3ae154341152799b31d8c00826c5caf1a7089505