BUG#49259: Slave I/O thread could not register on master
The slave thread changed the format of the information it used to
connect to the master after patch for BUG 13963.  This resulted
in old master getting confused, thence rejecting the slave
connection attempt.

In particular, patch for BUG 13963 removed the rpl_recovery_rank
variable which was, at that time, packed together with the rest
of the information which the slave would use to register itself
on the master. Based on this data, the master would then assert
that the number of bytes received in the connection command was
consistent to what it was expecting. 

Therefore, given that a slave, patched with the aforementioned
patch, would not pack the four bytes related to the
rpl_recovery_rank variable, the old master would reject the
connection attempt. It would assume that the data was
inconsistent (fewer bytes than it was expecting) and return 
an error.

We fix this by faking an rpl_recovery_rank variable when
registering the slave on the master. In practice this reverts a
small part of patch for BUG 13963, the one related to the slave
connecting to the master.
--BZR--
revision-id: luis.soares@sun.com-20091218025454-lv7rva4k82dwjze5
property-branch-nick: mysql-5.1-rep+3
property-file-info: ld7:file_id69:sp1f-repl_failsafe.cc-20011010025623-k7zhoyc3smc7tbliyp7vaf3f4idq22so7:message340:Added bypassing of removed rpl_recovery_rank variable information in
property-file-info: packet. This should also make more sense when old servers connect
property-file-info: to a new master (ie, master with patch for BUG 13963). If this was not
property-file-info: done, the new master could interpert information an old slave sends as
property-file-info: master_id, when in fact it could be the rpl_recovery_rank data.4:path20:sql/repl_failsafe.cced7:file_id61:sp1f-slave.cc-19700101030959-a636aj3mjxgu7fnznrg5kt77p3u2bvhh7:message79:Faking a rpl_recovery_rank so that we can register as a slave in
property-file-info: an old master.4:path12:sql/slave.ccee
testament3-sha1: ef32511c47d8406f0e421c9be09e13fec3771f1a