Bug #31145: ALTER TABLE DROP COLUMN, ADD COLUMN crashes (linux) or   freezes (win) the server
The check for equality was assuming the field object is always 
created. If it's not it was de-referencing a NULL pointer.
Fixed to use the data in the create object instead.
--BZR--
revision-id: joro@sun.com-20091218120030-zwz6r364ayxx5q83
property-branch-nick: B31145-5.1-bugteam
property-rebase-of: joro@sun.com-20091218120030-03vi5gbwodddlh58
testament3-sha1: fd1b36945d7821a90472799c2ab8154cc4247fa3