Fix for rpl_bug31076 valgrind failure which popped up after WL#5151 was pushed.
Problem 1: Some old binlog events do not contain metadata. This
makes checking whether the field can be converted or not rather
impossible because one cannot compare, for instance, field sizes
from original table and target table.

Solution 1: When an event does not contain metadata, we will just
check if field types are equal and assume that original field
definition matched with the one in the target table.

Problem 2: There is a second fix, which involves lack of
information regarding maybe_null. This was causing a conditional
jump warning when creating a conversion table. 

Solution 2: We will just assume that all fields that need to be
in the conversion table may be null.
--BZR--
revision-id: luis.soares@sun.com-20100107010438-07c3s3p6m0t1vf8a
property-branch-nick: mysql-5.1-rep+2-delivery1
testament3-sha1: 8c91800f630cbe6c06444c52e77951375b76a148