Fix for Bug#50221 Server refuses to start from non-default installdir on Solaris 32-bit
When starting mysqld it did not recognize most of the options given on
the command line when it was compiled for 32-bit Solaris using Sun
Studio compiler. The cause for this was that most of the entries in
the my_long_options array contained "garbage" data. The garbage data
was caused by a compiler bug. When initilizing the def_value member
for the "default-storage-engine" entry it was initialized like this:

  (longlong)"MyISAM"

i.e. casting a 32 bit pointer to a 64 bit integer value. Due to the
compiler bug only 4 bytes was allocated (instead of 8 bytes). This
caused everything following this entry to be stored at a location that
was 4 byte wrong.

The fix/work-around for this problem is initialize the def_value
for default-storage-engine in my_long_options to 0 and instead
initialize the default_storage_engine variable to "MyISAM" in
init_common_variables().
--BZR--
revision-id: olav@sun.com-20100114123942-ie1pbuub6qg6ll9n
property-branch-nick: bug50221-v2
property-file-info: ld7:file_id62:sp1f-mysqld.cc-19700101030959-zpswdvekpvixxzxf7gdtofzel7nywtfj7:message543:Due to a bug in Sun Studio compiler when generating 32 bit code the 
property-file-info: initialization of the def_value member of the default-storage-engine entry 
property-file-info: in my_long_options only got 4 bytes allocated instead of 8 bytes. 
property-file-info: The compiler bug was triggered by casting a 32 bit pointer to a 64 bit 
property-file-info: integer value in the initialization code for my_long_options. To avoid 
property-file-info: triggering the compiler bug the intialization of the def_value in
property-file-info: my_long_options is set to 0 and instead the default_storage_engine
property-file-info: is initialized to "MyISAM" in init_common_variables().4:path13:sql/mysqld.ccee
testament3-sha1: 92ec6c3b7b71cd1faa44c7a81525aa9349eed371