Bug#50409 Solaris 8 compatibility broken by assumption about printstack() being present
When Bug#47391 was fixed, no assumption was made that support
for Solaris 8 was needed. Solaris 8 lacks printstack(), and 
the build breaks because of this.

This patch adds a test for the presence of printstack() to
configure.in for 5.0, and uses HAVE_PRINTSTACK to make
decisions rather than the __sun define.

--BZR--
revision-id: staale.smedseng@sun.com-20100127103850-kk07g9x28hpzi89t
property-branch-nick: 50409-50
testament3-sha1: 5772d8e358267f3076a65ecb7d5cbbdc3223af85