branches/5.1: Undo the change from r6424. We need to return DB_SUCCESS even if we were unable to initialize the tabe autoinc value. This is required for the open to succeed. The only condition we currently treat as a hard error is if the autoinc field instance passed in by MySQL is NULL.
Previously if the table autoinc value was 0 and the next value was requested
we had an assertion that would fail. Change that assertion and treat a value
of 0 to mean that the autoinc system is unavailable. Generation of next
value will now return failure.

rb://237
--BZR--
revision-id: svn-v4:cee13dc7-1704-0410-992b-c9b4543f1246:branches/5.1:6535
testament3-sha1: b127437b1d6301f46b2650c697cd072b2c674777